Bump indicatif version from 0.17.2 to 0.18.3 (latest) (#24344)
## Description 

We are getting failing `cargo-deny`
[actions](https://github.com/MystenLabs/sui/actions/runs/19511548093/job/55852327532?pr=24283)
on this repo due to the `number_prefix` crate being marked unmaintained:

```
error[unmaintained]: number_prefix crate is unmaintained
    ┌─ /home/runner/work/sui/sui/Cargo.lock:774:1
    │
774 │ number_prefix 0.4.0 registry+https://github.com/rust-lang/crates.io-index
    │ ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ unmaintained advisory detected
    │
    ├ ID: RUSTSEC-2025-0119
    ├ Advisory: https://rustsec.org/advisories/RUSTSEC-2025-0119
    ├ The `number_prefix` crate is no longer actively maintained.  If you rely on this crate, consider switching to a recommended alternative.
      
      ## Recommended alternatives
      
      - [`unit-prefix`](https://crates.io/crates/unit-prefix)
    ├ Announcement: https://github.com/ogham/rust-number-prefix/pull/8
    ├ Solution: No safe upgrade is available!
```

This is a dep of `indicatif`. It's on `0.17.2` but the latest version is
`0.18.3`.

The latest version doesn't appear to use `number_prefix`, so bumping the
version should fix the issue.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: