[ci] Add windows ci for cli tests (#24357)
## Description 

This PR adds a job in the `rust.yml` workflow that runs tests for the
`sui` crate on Windows. We want to have this running with every PR as we
are planning for stronger Windows support in the near and long term.

In addition, it adds a linking flag for x64-86 msvc builds to increase
the stack to 8MB, to avoid overflowing the main thread stack (which
happens in debug mode for many sui commands).

## Test plan 

Existing tests should pass.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: