Certora audit fix: L-13 (#24285)
## Description 

This is a fix for issue L-13 from the Certora audit of the native
bridge.

The audit can be seen
[here](https://www.notion.so/mystenlabs/Certora-Audit-2966d9dcb4e980c490e2ee96b63dbe82?source=copy_link).

> Reqwest by default has no timeout. This can cause the request to hang
forever and keep a thread running (esp. in the eth client, which has no
other timeout). Therefore, it’s best to set a timeout for the client.

I also renamed `MeteredEthHttpProvier` to `MeteredEthHttpProvider`, as
that appears to be a typo.

In order to pass a `reqwest::Client` to
`ethers::providers::Http::new_with_client`, we need to have a matching
`reqwest` version with `ethers`. We were actually using `0.12.9`, but
`ethers` expects `0.11.20`. And the `ethers` repo is archived, with no
newer version. So to use `ethers` with a timeout we need to downgrade
`reqwest` to `0.11`. I implemented this, but we may want to keep the
current version instead, and just not bother passing a `Client` with a
timeout to ethers.

## Test plan 

Nothing other than CI tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: