[static ptb][hardening] check object type at the beginning of execution (#24376)
## Description 

- We will double check the type on disk against the type once we reach
execution. The type on disk is how we got it in the first place, but
this is just an added invariant check done at the beginning of
execution.
- We can safely remove this at anytime we feel comfortable

## Test plan 

- Ran tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: