[cli] Fix windows faucet wallet issue (#24384)
## Description 

This PR fixes an issue flagged in #24041. Unfortunately, Windows does
not allow clients to connect to 0.0.0.0, which was the case here, and
led to an error when the faucet had to make an RPC client to node.

## Test plan 

Existing tests. Manually tested on Windows!

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: