[indexer-alt-framework] Set checkpoint_hi_inclusive, reader_lo, pruner_hi if watermark does not exist (#24523)
## Description 

Without the change in this PR, the indexer logs these errors when
started using a non-zero `--first-checkpoint` for the first time (no
watermark record exists):
```
2025-12-03T20:23:45.594924Z ERROR sui_indexer_alt_framework::pipeline::concurrent::pruner: Failed to prune data for range: 121926000 to 121928000: No checkpoint mapping found for checkpoint 121926000 pipeline="kv_epoch_starts"
```

This error is caused by the pruner attempting to prune data that was
never indexed because `pruner_hi` is initialized to `0`.

This PR sets `checkpoint_hi_inclusive`, `reader_lo`, `pruner_hi` based
on the `default_next_checkpoint` which is either the value from
`--first-checkpoint` (defaulting to `0` if not set). When the pruner
runs, it uses this value of `pruner_hi` to avoid trying to prune data
that was never indexed.

## Test plan 

1. Added unit tests.
2. Tested with sui-indexer-alt-benchmark:
a. Deploy sui-indexer-alt-benchmark with `pulumi up` (this line causes
the benchmark to use the branch from this PR
https://github.com/MystenLabs/sui-operations/blob/c3011e3ae777c8b58019c4a83211aefa54f8d372/pulumi/gcp/sui-indexer-alt-benchmark/Pulumi.dev.yaml#L8)
  b. Verify error is not in logs 
    ```
kubectl logs sui-indexer-alt-benchmark-indexer-89f9c6965-s6hx9 -n
sui-indexer-alt-benchmark`
    ```
---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [x] Indexing Framework: Fix pruning for concurrent pipelines when
indexer is initialized with `--first-checkpoint`.